+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/command_rewrite/CommandRewriteInitializer.java
@@ -9,15 +9,25 @@ import io.github.sakurawald.fuji.core.document.annotation.TestCase;
 import io.github.sakurawald.fuji.core.structure.RegexRewriteNode;
 import io.github.sakurawald.fuji.module.initializer.ModuleInitializer;
 import io.github.sakurawald.fuji.module.initializer.command_rewrite.config.model.CommandRewriteConfigModel;
+import java.util.List;
 import org.jetbrains.annotations.NotNull;
 
 @Document(id = 1751826288031L, value = """
     This module allows you to define `regex` to rewrite the command line a player issued.
     """)
-@ColorBox(id = 1751971914915L, color = ColorBox.ColorBoxTypes.TIPS, value = """
-    ◉ You can use this module to shorten the command string.
-    Like, rewrite the command string `/home` into `/home tp default`.
-    And provide a shortcut, for `/home tp default` command.
+@ColorBox(id = 1756047806466L, color = ColorBox.ColorBoxTypes.NOTE, value = """
+    ◉ How it works?
+    1. Intercept the `issued command string` packets.
+    2. Apply the defined rewrite rules to `the command string` in top-down order.
+
+    <blue>NOTE: The `command_rewrite` occurs at the very beginning of the command lifecycle.
+    <blue>You will not receive `command suggestions` or `command exceptions` at this stage.
+    <blue>The rewrite rules are applied directly to the raw command string.
+
+    ◉ Command Lifecycle
+    1. **Issuing** – The command is sent as `a raw string`. At this stage, no command suggestions or exceptions are available.
+    2. **Parsing** – The command string is parsed into `a command node`. At this point, command suggestions and exceptions can be provided.
+    3. **Execution** – The `actions` associated with the command node are `executed`, producing the intended `effects` of the command.
     """)
 @ColorBox(id = 1751971980398L, color = ColorBox.ColorBoxTypes.NOTE, value = """
     ◉ The different compared to `command_alias` module and `command_bundle` module.
@@ -25,21 +35,34 @@ import org.jetbrains.annotations.NotNull;
     2. The `command bundle` module allows you to define `new command`, and define the argument and body of the new command.
     3. The `command rewrite` module didn't define any new command. It just works in `network packet` level, and modifies the `issued command string` from the command source player.
     """)
+@ColorBox(id = 1751971914915L, color = ColorBox.ColorBoxTypes.TIPS, value = """
+    ◉ You can use this module to shorten the command string.
+    Like, rewrite the command string `/home` into `/home tp default`.
+    And provide a shortcut, for `/home tp default` command.
+    """)
 @TestCase(action = "Issue `/home` command.", targets = "It should be rewrite to `/home tp default` command.")
 public class CommandRewriteInitializer extends ModuleInitializer {
     public static final BaseConfigurationHandler<CommandRewriteConfigModel> config = ObjectConfigurationHandler.ofModule(BaseConfigurationHandler.CONFIG_JSON_LITERAL, CommandRewriteConfigModel.class);
 
-    public static String rewriteCommand(@NotNull String oldString) {
-        for (RegexRewriteNode entry : CommandRewriteInitializer.config.model().rules) {
-            if (entry.getRegex() == null || entry.getReplacement() == null) {
-                LogUtil.warn("There is an invalid `null` entry in `command_rewrite.regex`, you should remove it: {}", entry);
-                continue;
-            }
+    public static @NotNull String processCommandRewrite(@NotNull String oldString) {
+        /* Compute effective rewrite rules. */
+        List<RegexRewriteNode> effectiveRewriteRules = CommandRewriteInitializer.config.model().rules
+            .stream()
+            .filter(rule -> {
+                boolean result = rule.getRegex() != null && rule.getReplacement() != null;
+                if (!result) {
+                    LogUtil.warn("Both regex property and replacement property should not be null, ignoring the command rewrite rule: {}", rule);
+                }
+                return result;
+            })
+            .toList();
 
-            if (oldString.matches(entry.getRegex())) {
-                String newString = oldString.replaceAll(entry.getRegex(), entry.getReplacement());
-                LogUtil.debug("Rewrite the command: old = {}, new = {}", oldString, newString);
-                return newString;
+        /* Applied the rewrite rules. */
+        for (RegexRewriteNode rewriteRule : effectiveRewriteRules) {
+            if (oldString.matches(rewriteRule.getRegex())) {
+                String newString = oldString.replaceAll(rewriteRule.getRegex(), rewriteRule.getReplacement());
+                LogUtil.debug("Rewrite the command string: old = {}, new = {}", oldString, newString);
+                oldString = newString;
             }
         }